+++ examples/next/CHANGELOG.md
@@ -1,5 +1,13 @@
 # example-next
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [2d92a8b]
+  - @brease/core@0.2.0
+  - @brease/react@0.1.2
+
 ## 0.1.2
 
 ### Patch Changes