+++ packages/react-brease-devtools/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @brease/react-devtools
 
+## 1.0.2
+
+### Patch Changes
+
+- @brease/react@0.1.2
+
 ## 1.0.1
 
 ### Patch Changes