+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brease/react
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [2d92a8b]
+  - @brease/core@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes